+++ frontend/src/pages/notes/Workspaces.js
@@ -15,7 +15,7 @@ import { Add, Close, MoreVert } from '@mui/icons-material';
 import Divider from '@mui/material/Divider';
 import { useNavigate } from 'react-router-dom';
 
-export default function Workspaces() {
+export default function Workspaces({ setAppBarHeader }) {
   // Misc
   const navigate = useNavigate();
 
@@ -43,8 +43,9 @@ export default function Workspaces() {
   }, [token]);
 
   useEffect(() => {
+    setAppBarHeader('Workspaces');
     fetchWorkspaces();
-  }, [fetchWorkspaces]);
+  }, [fetchWorkspaces, setAppBarHeader]);
 
   // Triple Dot Menu Functions
   const [tripleDotAnchorElement, setTripleDotAnchorElement] = useState(null);
@@ -164,11 +165,6 @@ export default function Workspaces() {
   return (
     <Container maxWidth="sm" sx={{ display: 'flex', justifyContent: 'center', alignItems: 'flex-start', py: 2 }}>
       <Paper elevation={3} sx={{ px: 1.5, py: 1.5, width: '100%', background:'var(--secondary-background-color)' }}>
-        {/* Header */}
-        <Typography variant="h4" align="center" gutterBottom sx={{ mt: 1.5, fontWeight: 'bold', color: 'var(--secondary-color)'}}>
-          Workspaces
-        </Typography>
-
         {/* This is for loading */}
         {loading && (
           <Typography align="center"> Loading… </Typography>
@@ -182,7 +178,6 @@ export default function Workspaces() {
         )}
 
         {/* If we're done loading and there are no errors */}
-        <Divider sx={{ borderBottomWidth: 2, marginBottom: 1, bgcolor: 'var(--secondary-color)' }} />
         {!loading && !error && (
           <Stack spacing={1}>
             {lists.length ? lists.map(list => (