+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project are documented in this file.
 ## 2025-08-11
 ### Added
 - Preliminary App Bar
+- Add dynamic appBarHeader updates across Workspaces, TodoLists, and Notes components
 ### Fixed
 - Filled in contact info placeholder in the `LICENSE.md`